Compression Function Main Loop Further Extended; Hash Values Allowed to Vary
The loop for the hash algorithm's main compression function was extended
to include the addition of the compressed chunk to the current hash
value. The hash values h_original in that area were renamed to h_new to
reflect the changing nature of those values each time the loop runs.
Accordingly, the h_new values were also initialized to the h_original
values at their declaration.