Minimization of Parallel Items
The declarations and initializations of the h_original and k constants
were moved into the main process block in order to minimize the running
of parallel items simultaneously (since Synthesize indicated that the
algorithm required more than 100% of the available resources). To
accomplish this, the values h_original and k were changed from constants
to variables.